refactor(web): split Settings page into submenus with left-rail nav (#1982)
