+++ apps/web/src/App.tsx
@@ -29,6 +29,11 @@ const PAGE_TITLES: Record<string, string> = {
   '/achievements': 'Achievements',
   '/watchlists': 'Watchlists',
   '/settings': 'Settings',
+  '/settings/account': 'Settings · Account',
+  '/settings/preferences': 'Settings · Preferences',
+  '/settings/privacy': 'Settings · Privacy & Data',
+  '/settings/sync': 'Settings · Sync & Devices',
+  '/settings/advanced': 'Settings · Advanced',
   '/import': 'Import',
   '/import/wizard': 'Import Wizard',
   '/import/receipt-ocr': 'Receipt OCR',
